# MAINT/ENH: stats.multivariate_normal.rvs: fix shape and speed up with Covariance