#### Reference issue
gh-16278
https://github.com/scipy/scipy/pull/12312#issuecomment-640298227 (direct call to BLAS can come later...)
supersedes gh-12184 

#### What does this implement/fix?
`stats.multivariate_normal.rvs` generates random variates using the Generator's `multivariate_normal`, which doesn't cache its factorization of the covariance matrix, so repeated calls require refactorizing the covariance matrix whether the distribution is frozen or not. The `Covariance` objects remember their factorization whether the distribution is frozen or not. 

Also, `stats.multivariate_normal.rvs` does an "indiscriminate squeeze", so singleton dimensions requested by the user are not present in the output. When the user passes an array as the covariance matrix, it is not easy to fix this in a backward-compatible way. But if the user passes a new `Covariance` instance, it might as well return an array of the correct shape.

This PR adds a `colorize` method to the `Covariance` class and uses it to generate random variates in `multivariate_normal.rvs`.

```python3
import numpy as np
from scipy import stats
import matplotlib.pyplot as plt

rng = np.random.default_rng(984261616848)
size = (10, 10, 10)

mean = [10, 20]
V = np.array([[1, 1], [-1, 1]]) * np.sqrt(2)/2
W = np.diag([1, 5])
A = V.T @ W @ V

rvs1 = stats.multivariate_normal.rvs(mean, A, size=size, random_state=rng)
rvs1 = rvs1.reshape(1000, -1)

cov = stats.Covariance.from_cholesky(np.linalg.cholesky(A))
# cov = stats.Covariance.from_precision(np.linalg.inv(A))
# cov = stats.Covariance.from_eigendecomposition(np.linalg.eigh(A))
rvs2 = stats.multivariate_normal.rvs(mean, cov, size=size, random_state=rng)
rvs2 = rvs2.reshape(1000, -1)

plt.plot(*rvs1.T, '.', *rvs2.T, '.', alpha=0.5)
# plt.plot(*rvs2.T, '.', *rvs1.T, '.', alpha=0.5)  # looks the same
```

![image](https://user-images.githubusercontent.com/6570539/197363116-e3d5412d-900f-484c-9a3a-2d5f4031f1a7.png)

<details>

Even in low dimensional spaces and considering the cost of `Covariance` construction, the speed of the `Covariance` versions is comparable to that of NumPy's generator, which we currently rely on.

```python3
import numpy as np
from scipy import stats
rng = np.random.default_rng(984261616848)

mean = rng.random(10)
A = np.random.rand(10, 10)
A = A @ A.T

# One sample
%timeit rng.multivariate_normal(mean, A, size=None)
87 µs ± 811 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit cov = stats.Covariance.from_cholesky(np.linalg.cholesky(A)); stats.multivariate_normal.rvs(mean, cov, size=None, random_state=rng)
# 45.8 µs ± 107 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit cov = stats.Covariance.from_eigendecomposition(np.linalg.eigh(A)); stats.multivariate_normal.rvs(mean, cov, size=None, random_state=rng)
# 116 µs ± 1.2 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)

# Many samples
size = (10, 10, 10)

%timeit rng.multivariate_normal(mean, A, size=size)
# 207 µs ± 7.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

%timeit cov = stats.Covariance.from_cholesky(np.linalg.cholesky(A)); stats.multivariate_normal.rvs(mean, cov, size=size, random_state=rng)
# 205 µs ± 734 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

%timeit cov = stats.Covariance.from_eigendecomposition(np.linalg.eigh(A)); stats.multivariate_normal.rvs(mean, cov, size=size, random_state=rng)
# 292 µs ± 419 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

The situation is even better in high-dimensional spaces.
```python3
n = 1000
mean = rng.random(n)
A = np.random.rand(n, n)
A = A @ A.T

%timeit rng.multivariate_normal(mean, A, size=None)
# 165 ms ± 503 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit cov = stats.Covariance.from_cholesky(np.linalg.cholesky(A)); stats.multivariate_normal.rvs(mean, cov, size=None, random_state=rng)
# 16.3 ms ± 92.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

%timeit cov = stats.Covariance.from_eigendecomposition(np.linalg.eigh(A)); stats.multivariate_normal.rvs(mean, cov, size=None, random_state=rng)
# 59.5 ms ± 1.81 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

And of course, if the covariance matrix is diagonal, the `Covariance` object created by `from_diagonal` is more accurate and faster still.

</details>


#### Additional information
Should the method be called `color` or `colorize`?

We'll fix the shape of `pdf`/`logpdf` output in a separate PR. We can use `Covariance` similarly in a few other multivariate distributions. IIUC, part of the speed increase is because these never use SVD. If we need a `from_SVD` version, it can easily be added.

@siddhantwahal