@tirthasheshpatel 2022-10-23
> Alternatively, if CI won't complain about this lint issue after it appears in `main`, we could just merge as-is.

That sounds good. Main doesn't run the lint check so should not be a problem.